+++ Invoice Generator/src/com/bill/edittoaddress/EditToAddressController.java
@@ -1,6 +1,7 @@
 package com.bill.edittoaddress;
 
 import java.net.URL;
+import java.sql.SQLException;
 import java.util.ResourceBundle;
 
 import com.bill.beans.Address;
@@ -106,7 +107,8 @@ public class EditToAddressController implements Initializable {
 		try {
 			if(ValidateUserInputs.validateAddressDetails(updateOldShopName.getValue(), updateNewShopName.getText(), updateAddressLine1.getText(),
 					updateAddressLine2.getText(), updateCity.getText(), updateDistrict.getText(), updateState.getText(), 
-					updatePincode.getText(), updateTelephone.getText(), updateMobile.getText(), updateGstNo.getText())) {
+					updatePincode.getText(), updateTelephone.getText(), updateMobile.getText(), updateGstNo.getText())
+					&& ShowPopups.showPopups(AlertType.CONFIRMATION, "Are you sure want to update this 'Bill To' address..", "")) {
 				
 				ToDatabaseValidator.updateToAddress(updateOldShopName.getValue(), updateNewShopName.getText(), updateAddressLine1.getText(),
 						updateAddressLine2.getText(), updateCity.getText(), updateDistrict.getText(), updateState.getText(), 
@@ -114,6 +116,11 @@ public class EditToAddressController implements Initializable {
 				
 				ShowPopups.showPopups(AlertType.INFORMATION, "Success....", "Address Updated Successfully....");
 			}
+		} catch (SQLException e) {
+			if(e.getMessage().contains("UNIQUE"))
+				ShowPopups.showPopups(AlertType.ERROR, "Shop name is same as the existing one. Please provide different shop name", "");
+			else
+				ShowPopups.showPopups(AlertType.ERROR, e.toString(), "");
 		} catch (Exception e) {
 			ShowPopups.showPopups(AlertType.ERROR, e.toString(), "");
 		}
@@ -125,7 +132,8 @@ public class EditToAddressController implements Initializable {
 		try {
 			if(ValidateUserInputs.validateAddressDetails(insertShopName.getText(),insertShopName.getText(), insertAddressLine1.getText(),
 					insertAddressLine2.getText(), insertCity.getText(), insertDistrict.getText(), insertState.getText(), 
-					insertPincode.getText(), insertTelephone.getText(), insertMobile.getText(), insertGstNo.getText())) {
+					insertPincode.getText(), insertTelephone.getText(), insertMobile.getText(), insertGstNo.getText())
+					&& ShowPopups.showPopups(AlertType.CONFIRMATION, "Are you sure want to insert this 'Bill To' address..", "")) {
 				
 				ToDatabaseValidator.insertToAddress(insertShopName.getText(), insertAddressLine1.getText(),
 						insertAddressLine2.getText(), insertCity.getText(), insertDistrict.getText(), insertState.getText(), 
@@ -134,6 +142,11 @@ public class EditToAddressController implements Initializable {
 				ShowPopups.showPopups(AlertType.INFORMATION, "Success....", "Address Inserted to Database Successfully....");
 
 			}
+		} catch (SQLException e) {
+			if(e.getMessage().contains("UNIQUE"))
+				ShowPopups.showPopups(AlertType.ERROR, "Shop name is same as the existing one. Please provide different shop name", "");
+			else
+				ShowPopups.showPopups(AlertType.ERROR, e.toString(), "");
 		} catch (Exception e) {
 			ShowPopups.showPopups(AlertType.ERROR, e.toString(), "");
 		}
@@ -143,7 +156,8 @@ public class EditToAddressController implements Initializable {
 	public void deleteAddress() {
 		
 		try {
-			if(ValidateUserInputs.validateShopName(deleteShopName.getValue())) {
+			if(ValidateUserInputs.validateShopName(deleteShopName.getValue())
+					&& ShowPopups.showPopups(AlertType.CONFIRMATION, "Are you sure want to delete this 'Bill To' address..", "")) {
 				
 				ToDatabaseValidator.deleteToAddress(deleteShopName.getValue());
 				ShowPopups.showPopups(AlertType.INFORMATION, "Success....", "Address Deleted from the Database Successfully....");
@@ -151,5 +165,21 @@ public class EditToAddressController implements Initializable {
 		} catch (Exception e) {
 			ShowPopups.showPopups(AlertType.ERROR, e.toString(), "");
 		}
+	}
+	
+
+	@FXML
+	public void disableFields() {
+		updateNewShopName.setDisable(true);
+		updateAddressLine1.setDisable(true);
+		updateAddressLine2.setDisable(true);
+		updateCity.setDisable(true);
+		updateDistrict.setDisable(true);
+		updateState.setDisable(true);
+		updatePincode.setDisable(true);
+		updateTelephone.setDisable(true);
+		updateMobile.setDisable(true);
+		updateGstNo.setDisable(true);
+		updateAddressBtn.setDisable(true);
 	}
 }
